rev 2 - competition list component are made composable (for real this time)
